Recognise umlauts typed as -e on non DE keyboard
This allows Germans using an English keyboard to type the Umlauts as "ue" "ae" and "oe" as is customary, and type the ß as "ss". 
Previously the spellchecker would not recognise these spellings.
As a test, try the word "Greetings"  as "Gruesse". With this fix, the correct intended spelling "Grüße" is given as a suggestion.